Add centroid qualifier to lightmap uv vertex shader output
This fixes lightmap artifacts with MSAA on Intel Cherry Trail.
It also seems to be a deviation from the GLSL 4.3 spec:

"When interpolation or auxiliary qualifiers do not match,
those provided in the fragment shader supersede those
provided in previous stages. [...] That is, what matters
is what is declared in the fragment shaders, not what is
declared in shaders in previous stages."

The OpenGL Shading Language 4.3, Section 4.3.4 (Input Variables), page 45
https://www.khronos.org/registry/OpenGL/specs/gl/GLSLangSpec.4.30.pdf#page=51